Clean up of the bot commands structure
