+++ main.go
@@ -7,13 +7,8 @@ package main
 import (
 	"bufio"
 	"context"
-	"encoding/base64"
-	"encoding/json"
 	"flag"
 	"fmt"
-	"io"
-	"net/http"
-	"net/url"
 	"os"
 	"os/signal"
 	"path/filepath"
@@ -24,11 +19,11 @@ import (
 	"github.com/companyzero/bisonrelay/clientrpc/types"
 	"github.com/companyzero/bisonrelay/zkidentity"
 	"github.com/decred/dcrd/dcrutil/v4"
-	"github.com/karamble/braibot/internal/audio"
+	"github.com/karamble/braibot/internal/commands"
+	braiconfig "github.com/karamble/braibot/internal/config"
 	"github.com/karamble/braibot/internal/database"
-	"github.com/karamble/braibot/internal/falapi"
 	kit "github.com/vctt94/bisonbotkit"
-	"github.com/vctt94/bisonbotkit/config"
+	botkitconfig "github.com/vctt94/bisonbotkit/config"
 	"github.com/vctt94/bisonbotkit/logging"
 	"github.com/vctt94/bisonbotkit/utils"
 )
@@ -38,501 +33,6 @@ var (
 	debug       = true              // Set to true for debugging
 	dbManager   *database.DBManager // Database manager for user balances
 )
-
-// Command represents a bot command
-type Command struct {
-	Name        string
-	Description string
-	Handler     func(ctx context.Context, bot *kit.Bot, cfg *config.BotConfig, pm types.ReceivedPM, args []string) error
-}
-
-// Available commands
-var commands map[string]Command
-
-func init() {
-	commands = map[string]Command{
-		"help": {
-			Name:        "help",
-			Description: "Shows this help message",
-			Handler: func(ctx context.Context, bot *kit.Bot, cfg *config.BotConfig, pm types.ReceivedPM, args []string) error {
-				helpMsg := "| Command | Description |\n| -------- | ----------- |\n"
-				for _, cmd := range commands {
-					helpMsg += fmt.Sprintf("| !%s | %s |\n", cmd.Name, cmd.Description)
-				}
-				return bot.SendPM(ctx, pm.Nick, helpMsg)
-			},
-		},
-		"listmodels": {
-			Name:        "listmodels",
-			Description: "Lists available models for a specific command. Usage: !listmodels [command]",
-			Handler: func(ctx context.Context, bot *kit.Bot, cfg *config.BotConfig, pm types.ReceivedPM, args []string) error {
-				if len(args) == 0 {
-					return bot.SendPM(ctx, pm.Nick, "Please specify a command (text2image or text2speech). Usage: !listmodels [command]")
-				}
-
-				commandName := strings.ToLower(args[0])
-
-				var modelList string
-				var models map[string]falapi.Model
-
-				switch commandName {
-				case "text2image":
-					models = falapi.Text2ImageModels
-					modelList = "Available models for text2image:\n| Model | Description | Price |\n| ----- | ----------- | ----- |\n"
-				case "text2speech":
-					models = falapi.Text2SpeechModels
-					modelList = "Available models for text2speech:\n| Model | Description | Price |\n| ----- | ----------- | ----- |\n"
-				default:
-					return bot.SendPM(ctx, pm.Nick, "Invalid command. Use 'text2image' or 'text2speech'.")
-				}
-
-				for _, model := range models {
-					modelList += fmt.Sprintf("| %s | %s | $%.2f |\n", model.Name, model.Description, model.Price)
-				}
-
-				return bot.SendPM(ctx, pm.Nick, modelList)
-			},
-		},
-		"setmodel": {
-			Name:        "setmodel",
-			Description: "Sets the model to use for specified commands. Usage: !setmodel [command] [modelname]",
-			Handler: func(ctx context.Context, bot *kit.Bot, cfg *config.BotConfig, pm types.ReceivedPM, args []string) error {
-				if len(args) < 2 {
-					return bot.SendPM(ctx, pm.Nick, "Please specify a command and a model name. Usage: !setmodel [command] [modelname]")
-				}
-				commandName := args[0]
-				modelName := args[1]
-
-				// Check if the command is valid
-				if _, exists := commands[commandName]; !exists {
-					return bot.SendPM(ctx, pm.Nick, "Invalid command name. Use !listmodels to see available commands.")
-				}
-
-				// Check if the model is valid for the specific command
-				var models map[string]falapi.Model
-				switch commandName {
-				case "text2image":
-					models = falapi.Text2ImageModels
-				case "text2speech":
-					models = falapi.Text2SpeechModels
-				default:
-					return bot.SendPM(ctx, pm.Nick, "Invalid command. Use 'text2image' or 'text2speech'.")
-				}
-
-				if _, exists := models[modelName]; exists {
-					falapi.SetDefaultModel(commandName, modelName)
-					return bot.SendPM(ctx, pm.Nick, fmt.Sprintf("Model for %s set to: %s", commandName, modelName))
-				}
-
-				return bot.SendPM(ctx, pm.Nick, fmt.Sprintf("Invalid model name for %s. Use !listmodels %s to see available models.", commandName, commandName))
-			},
-		},
-		"text2image": {
-			Name:        "text2image",
-			Description: "Generates an image from text prompt. Usage: !text2image [prompt]",
-			Handler: func(ctx context.Context, bot *kit.Bot, cfg *config.BotConfig, pm types.ReceivedPM, args []string) error {
-				if len(args) == 0 {
-					return bot.SendPM(ctx, pm.Nick, "Please provide a prompt. Usage: !text2image [prompt]")
-				}
-
-				prompt := strings.Join(args, " ")
-
-				// Create Fal.ai client
-				client := falapi.NewClient(cfg.ExtraConfig["falapikey"], debug)
-
-				// Get model configuration
-				modelName, exists := falapi.GetDefaultModel("text2image")
-				if !exists {
-					return fmt.Errorf("no default model found for text2image")
-				}
-				model, exists := falapi.GetModel(modelName, "text2image")
-				if !exists {
-					return fmt.Errorf("model not found: %s", modelName)
-				}
-
-				// Generate image
-				imageResp, err := client.GenerateImage(ctx, prompt, model.Name, bot, pm.Nick)
-				if err != nil {
-					return err
-				}
-
-				// Assuming the first image is the one we want to send
-				if len(imageResp.Images) > 0 {
-					imageURL := imageResp.Images[0].URL
-					// Fetch the image data
-					imgResp, err := http.Get(imageURL)
-					if err != nil {
-						return err
-					}
-					defer imgResp.Body.Close()
-
-					imgData, err := io.ReadAll(imgResp.Body)
-					if err != nil {
-						return err
-					}
-
-					// Encode the image data to base64
-					encodedImage := base64.StdEncoding.EncodeToString(imgData)
-
-					// Determine the image type from ContentType
-					var imageType string
-					switch imageResp.Images[0].ContentType {
-					case "image/jpeg":
-						imageType = "image/jpeg"
-					case "image/png":
-						imageType = "image/png"
-					case "image/webp":
-						imageType = "image/webp"
-					default:
-						imageType = "image/jpeg" // Fallback to jpeg if unknown
-					}
-
-					// Create the message with embedded image, using the user's prompt as the alt text
-					message := fmt.Sprintf("--embed[alt=%s,type=%s,data=%s]--", url.QueryEscape(prompt), imageType, encodedImage)
-					return bot.SendPM(ctx, pm.Nick, message)
-				} else {
-					return bot.SendPM(ctx, pm.Nick, "No images were generated.")
-				}
-			},
-		},
-		"balance": {
-			Name:        "balance",
-			Description: "Shows your current balance",
-			Handler: func(ctx context.Context, bot *kit.Bot, cfg *config.BotConfig, pm types.ReceivedPM, args []string) error {
-				// Convert UID to string ID for database, just like in tip handler
-				var userID zkidentity.ShortID
-				userID.FromBytes(pm.Uid)
-				userIDStr := userID.String()
-
-				// Get balance from database using the proper ID
-				balance, err := dbManager.GetBalance(userIDStr)
-				if err != nil {
-					return fmt.Errorf("failed to get balance: %v", err)
-				}
-
-				// Convert to DCR using dcrutil, same as in tip handler
-				dcrBalance := dcrutil.Amount(balance / 1e3).ToCoin()
-
-				// Send balance message
-				return bot.SendPM(ctx, pm.Nick, fmt.Sprintf("Your current balance is %.8f DCR", dcrBalance))
-			},
-		},
-		"rate": {
-			Name:        "rate",
-			Description: "Shows current DCR exchange rate in USD and BTC",
-			Handler: func(ctx context.Context, bot *kit.Bot, cfg *config.BotConfig, pm types.ReceivedPM, args []string) error {
-				// Send a status message to indicate we're fetching rates
-				bot.SendPM(ctx, pm.Nick, "Fetching current exchange rates...")
-
-				// Create HTTP client with timeout
-				client := &http.Client{
-					Timeout: 10 * time.Second,
-				}
-
-				// Make request to dcrdata API
-				resp, err := client.Get("https://explorer.dcrdata.org/api/exchangerate")
-				if err != nil {
-					return bot.SendPM(ctx, pm.Nick, fmt.Sprintf("Error fetching rates: %v", err))
-				}
-				defer resp.Body.Close()
-
-				// Check status code
-				if resp.StatusCode != http.StatusOK {
-					return bot.SendPM(ctx, pm.Nick, fmt.Sprintf("Error: API returned status %d", resp.StatusCode))
-				}
-
-				var rates struct {
-					DCRPrice float64 `json:"dcrPrice"`
-					BTCPrice float64 `json:"btcPrice"`
-					Time     int64   `json:"time"`
-				}
-
-				if err := json.NewDecoder(resp.Body).Decode(&rates); err != nil {
-					return bot.SendPM(ctx, pm.Nick, fmt.Sprintf("Error parsing rates: %v", err))
-				}
-
-				// Convert timestamp to human readable format
-				timeStr := time.Unix(rates.Time, 0).Format(time.RFC1123)
-
-				// Format the response with more detailed information
-				message := fmt.Sprintf("Current DCR Exchange Rates (as of %s):\n"+
-					"USD: $%.2f\n"+
-					"BTC: %.8f BTC\n"+
-					"Source: dcrdata",
-					timeStr, rates.DCRPrice, rates.BTCPrice)
-
-				return bot.SendPM(ctx, pm.Nick, message)
-			},
-		},
-		"text2speech": {
-			Name:        "text2speech",
-			Description: "Converts text to speech. Usage: !text2speech [voice_id] [text] - voice_id is optional, defaults to Wise_Woman. Available voices: Wise_Woman, Friendly_Person, Inspirational_girl, Deep_Voice_Man, Calm_Woman, Casual_Guy, Lively_Girl, Patient_Man, Young_Knight, Determined_Man, Lovely_Girl, Decent_Boy, Imposing_Manner, Elegant_Man, Abbess, Sweet_Girl_2, Exuberant_Girl",
-			Handler: func(ctx context.Context, bot *kit.Bot, cfg *config.BotConfig, pm types.ReceivedPM, args []string) error {
-				if len(args) < 2 {
-					voiceList := "Available voices: Wise_Woman, Friendly_Person, Inspirational_girl, Deep_Voice_Man, Calm_Woman, Casual_Guy, Lively_Girl, Patient_Man, Young_Knight, Determined_Man, Lovely_Girl, Decent_Boy, Imposing_Manner, Elegant_Man, Abbess, Sweet_Girl_2, Exuberant_Girl"
-					return bot.SendPM(ctx, pm.Nick, fmt.Sprintf("Please provide a voice ID and text. Usage: !text2speech [voice_id] [text]\n\n%s", voiceList))
-				}
-
-				voiceID := args[0]
-				text := strings.Join(args[1:], " ")
-
-				// Create Fal.ai client
-				client := falapi.NewClient(cfg.ExtraConfig["falapikey"], debug)
-
-				// Get model configuration
-				modelName, exists := falapi.GetDefaultModel("text2speech")
-				if !exists {
-					return fmt.Errorf("no default model found for text2speech")
-				}
-				_, exists = falapi.GetModel(modelName, "text2speech")
-				if !exists {
-					return fmt.Errorf("model not found: %s", modelName)
-				}
-
-				// Generate speech
-				audioResp, err := client.GenerateSpeech(ctx, text, voiceID, bot, pm.Nick)
-				if err != nil {
-					return err
-				}
-
-				// Fetch the audio data
-				resp, err := http.Get(audioResp.Audio.URL)
-				if err != nil {
-					return err
-				}
-				defer resp.Body.Close()
-
-				audioData, err := io.ReadAll(resp.Body)
-				if err != nil {
-					return err
-				}
-
-				// Convert PCM to Opus using the audio package
-				oggData, err := audio.ConvertPCMToOpus(audioData)
-				if err != nil {
-					return err
-				}
-
-				// Encode the audio data to base64
-				encodedAudio := base64.StdEncoding.EncodeToString(oggData)
-
-				// Create the message with embedded audio
-				message := fmt.Sprintf("--embed[alt=%s,type=audio/ogg,data=%s]--", url.QueryEscape(text), encodedAudio)
-				return bot.SendPM(ctx, pm.Nick, message)
-			},
-		},
-	}
-}
-
-// isCommand checks if a message is a command (starts with !)
-func isCommand(msg string) (string, []string, bool) {
-	if !strings.HasPrefix(msg, "!") {
-		return "", nil, false
-	}
-
-	parts := strings.Fields(msg[1:]) // Remove ! and split
-	if len(parts) == 0 {
-		return "", nil, false
-	}
-
-	cmd := strings.ToLower(parts[0])
-	args := parts[1:]
-	return cmd, args, true
-}
-
-// testAssetServerCredentials tests if the asset server credentials are valid
-func testAssetServerCredentials(serverURL, apiKey string) error {
-	// Create HTTP client with timeout
-	client := &http.Client{
-		Timeout: 10 * time.Second,
-	}
-
-	// Create request
-	req, err := http.NewRequest("GET", serverURL+"/test", nil)
-	if err != nil {
-		return fmt.Errorf("error creating request: %v", err)
-	}
-
-	// Add API key header
-	req.Header.Set("X-API-Key", apiKey)
-
-	// Send request
-	resp, err := client.Do(req)
-	if err != nil {
-		return fmt.Errorf("error connecting to asset server: %v", err)
-	}
-	defer resp.Body.Close()
-
-	// Check response status
-	if resp.StatusCode != http.StatusOK {
-		return fmt.Errorf("asset server returned error status %d", resp.StatusCode)
-	}
-
-	// Parse response
-	var result struct {
-		Success bool   `json:"success"`
-		Message string `json:"message"`
-	}
-	if err := json.NewDecoder(resp.Body).Decode(&result); err != nil {
-		return fmt.Errorf("error parsing response: %v", err)
-	}
-
-	if !result.Success {
-		return fmt.Errorf("asset server test failed: %s", result.Message)
-	}
-
-	return nil
-}
-
-// checkAndUpdateConfig checks if required configuration settings are present
-// and prompts the user to enter them if they're missing.
-func checkAndUpdateConfig(cfg *config.BotConfig, appRoot string) error {
-	// Ensure the directory exists first
-	if err := os.MkdirAll(appRoot, 0755); err != nil {
-		return fmt.Errorf("error creating app root directory: %v", err)
-	}
-
-	// Check for falapikey
-	if _, exists := cfg.ExtraConfig["falapikey"]; !exists {
-		fmt.Print("Enter your Fal.ai API key: ")
-		reader := bufio.NewReader(os.Stdin)
-		apiKey, err := reader.ReadString('\n')
-		if err != nil {
-			return fmt.Errorf("error reading API key: %v", err)
-		}
-		apiKey = strings.TrimSpace(apiKey)
-		if apiKey == "" {
-			return fmt.Errorf("API key cannot be empty")
-		}
-
-		// Update config in memory
-		cfg.ExtraConfig["falapikey"] = apiKey
-
-		// Append to config file
-		configPath := filepath.Join(appRoot, "braibot.conf")
-		f, err := os.OpenFile(configPath, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
-		if err != nil {
-			return fmt.Errorf("error opening config file: %v", err)
-		}
-		defer f.Close()
-
-		if _, err := f.WriteString(fmt.Sprintf("falapikey=%s\n", apiKey)); err != nil {
-			return fmt.Errorf("error writing to config file: %v", err)
-		}
-	}
-
-	// Check for asset server settings
-	if _, exists := cfg.ExtraConfig["use_assetserver"]; !exists {
-		var useAssetServer string
-		reader := bufio.NewReader(os.Stdin)
-		for {
-			fmt.Print("Do you want to use the asset server? (y/N): ")
-			input, err := reader.ReadString('\n')
-			if err != nil {
-				return fmt.Errorf("error reading asset server preference: %v", err)
-			}
-			useAssetServer = strings.TrimSpace(strings.ToLower(input))
-			if useAssetServer == "y" || useAssetServer == "n" || useAssetServer == "" {
-				break
-			}
-			fmt.Println("Please enter 'y' or 'n' (or press Enter for no)")
-		}
-
-		// Convert y/n to true/false, default to false if empty
-		useAssetServerValue := "false"
-		if useAssetServer == "y" {
-			useAssetServerValue = "true"
-		}
-
-		// Update config in memory
-		cfg.ExtraConfig["use_assetserver"] = useAssetServerValue
-
-		// Append to config file
-		configPath := filepath.Join(appRoot, "braibot.conf")
-		f, err := os.OpenFile(configPath, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
-		if err != nil {
-			return fmt.Errorf("error opening config file: %v", err)
-		}
-		defer f.Close()
-
-		if _, err := f.WriteString(fmt.Sprintf("use_assetserver=%s\n", useAssetServerValue)); err != nil {
-			return fmt.Errorf("error writing to config file: %v", err)
-		}
-
-		// If asset server is enabled, get additional settings
-		if useAssetServerValue == "true" {
-			// Get asset server URL
-			fmt.Print("Enter the asset server URL (e.g., https://assets.example.com): ")
-			assetServerURL, err := reader.ReadString('\n')
-			if err != nil {
-				return fmt.Errorf("error reading asset server URL: %v", err)
-			}
-			assetServerURL = strings.TrimSpace(assetServerURL)
-			if assetServerURL == "" {
-				return fmt.Errorf("asset server URL cannot be empty")
-			}
-
-			// Validate URL format
-			if !strings.HasPrefix(assetServerURL, "https://") {
-				return fmt.Errorf("asset server URL must start with https://")
-			}
-
-			// Update config in memory
-			cfg.ExtraConfig["assetserver_url"] = assetServerURL
-
-			// Append to config file
-			if _, err := f.WriteString(fmt.Sprintf("assetserver_url=%s\n", assetServerURL)); err != nil {
-				return fmt.Errorf("error writing to config file: %v", err)
-			}
-
-			// Get API secret
-			fmt.Print("Enter the asset server API secret: ")
-			apiSecret, err := reader.ReadString('\n')
-			if err != nil {
-				return fmt.Errorf("error reading API secret: %v", err)
-			}
-			apiSecret = strings.TrimSpace(apiSecret)
-			if apiSecret == "" {
-				return fmt.Errorf("API secret cannot be empty")
-			}
-
-			// Update config in memory
-			cfg.ExtraConfig["assetserver_secret"] = apiSecret
-
-			// Append to config file
-			if _, err := f.WriteString(fmt.Sprintf("assetserver_secret=%s\n", apiSecret)); err != nil {
-				return fmt.Errorf("error writing to config file: %v", err)
-			}
-
-			// Test the credentials
-			fmt.Println("Testing asset server credentials...")
-			if err := testAssetServerCredentials(assetServerURL, apiSecret); err != nil {
-				fmt.Printf("Asset server test failed: %v\n", err)
-				fmt.Println("Asset server will be disabled. You can try again later by editing the config file.")
-
-				// Set use_assetserver to false
-				useAssetServerValue = "false"
-				cfg.ExtraConfig["use_assetserver"] = "false"
-
-				// Update the config file with the new value
-				configPath := filepath.Join(appRoot, "braibot.conf")
-				f, err := os.OpenFile(configPath, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
-				if err != nil {
-					return fmt.Errorf("error opening config file: %v", err)
-				}
-				defer f.Close()
-
-				if _, err := f.WriteString(fmt.Sprintf("use_assetserver=false\n")); err != nil {
-					return fmt.Errorf("error writing to config file: %v", err)
-				}
-			} else {
-				fmt.Println("Asset server credentials verified successfully!")
-			}
-		}
-	}
-
-	return nil
-}
 
 func realMain() error {
 	flag.Parse()
@@ -564,7 +64,7 @@ func realMain() error {
 	log := logBackend.Logger("BraiBot")
 
 	// Load bot configuration
-	cfg, err := config.LoadBotConfig(appRoot, "braibot.conf")
+	cfg, err := botkitconfig.LoadBotConfig(appRoot, "braibot.conf")
 	if err != nil {
 		return fmt.Errorf("failed to load config: %v", err)
 	}
@@ -573,8 +73,8 @@ func realMain() error {
 	configPath := filepath.Join(appRoot, "braibot.conf")
 	for i := 0; i < 5; i++ { // Try for 5 seconds
 		if _, err := os.Stat(configPath); err == nil {
-			// File exists, proceed with checkAndUpdateConfig
-			if err := checkAndUpdateConfig(cfg, appRoot); err != nil {
+			// File exists, proceed with checkAndUpdateConfig for braibot specific configuration fields
+			if err := braiconfig.CheckAndUpdateConfig(cfg, appRoot); err != nil {
 				return fmt.Errorf("failed to check and update config: %v", err)
 			}
 			break
@@ -603,14 +103,17 @@ func realMain() error {
 		return fmt.Errorf("failed to create bot: %v", err)
 	}
 
+	// Initialize command registry
+	commandRegistry := commands.InitializeCommands(dbManager, debug)
+
 	// Add a goroutine to handle PMs using our bidirectional channel
 	go func() {
 		for pm := range pmChan {
 			log.Infof("Received PM from %s: %s", pm.Nick, pm.Msg.Message)
 
 			// Check if the message is a command
-			if cmd, args, isCmd := isCommand(pm.Msg.Message); isCmd {
-				if command, exists := commands[cmd]; exists {
+			if cmd, args, isCmd := commands.IsCommand(pm.Msg.Message); isCmd {
+				if command, exists := commandRegistry.Get(cmd); exists {
 					if err := command.Handler(context.Background(), bot, cfg, pm, args); err != nil {
 						log.Warnf("Error executing command %s: %v", cmd, err)
 					}
